rename: 나의 방 myroom으로 클래스명 변경
